+++ pkg/plugin/rbac.go
@@ -3,12 +3,13 @@ package plugin
 import (
 	"fmt"
 	"github.com/pkg/errors"
+	"github.com/robertsmieja/kubectl-purge/pkg/util"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"sync"
 )
 
-func deleteRoles(clientset *kubernetes.Clientset, namespace string, errorCh chan<- error) {
+func deleteRoles(clientset *kubernetes.Clientset, namespace string, logCh chan<- string, errorCh chan<- error) {
 	ctx, cancel := createCtx()
 	waitGroup := sync.WaitGroup{}
 
@@ -34,7 +35,7 @@ func deleteRoles(clientset *kubernetes.Clientset, namespace string, errorCh chan
 	waitGroup.Wait()
 }
 
-func deleteRoleBindings(clientset *kubernetes.Clientset, namespace string, errorCh chan<- error) {
+func deleteRoleBindings(clientset *kubernetes.Clientset, namespace string, logCh chan<- string, errorCh chan<- error) {
 	ctx, cancel := createCtx()
 	waitGroup := sync.WaitGroup{}
 
@@ -60,7 +61,20 @@ func deleteRoleBindings(clientset *kubernetes.Clientset, namespace string, error
 	waitGroup.Wait()
 }
 
-func deleteClusterRoles(clientset *kubernetes.Clientset, errorCh chan<- error) {
+var defaultClusterRoles = []string{
+	"admin",
+	"cluster-admin",
+	"edit",
+	"view",
+	"vpnkit-controller", // docker desktop
+}
+
+var defaultClusterRolePrefixes = []string{
+	"kubeadm:",
+	"system:",
+}
+
+func deleteClusterRoles(clientset *kubernetes.Clientset, logCh chan<- string, errorCh chan<- error) {
 	ctx, cancel := createCtx()
 	waitGroup := sync.WaitGroup{}
 
@@ -71,12 +85,16 @@ func deleteClusterRoles(clientset *kubernetes.Clientset, errorCh chan<- error) {
 		errorCh <- errors.Wrap(err, "failed to list clusterRoles")
 	}
 	for _, clusterRole := range clusterRoles.Items {
-		waitGroup.Add(1)
-
 		name := clusterRole.Name
+
+		if util.Contains(defaultClusterRoles, name) || util.StartsWithAny(defaultClusterRolePrefixes, name) {
+			logCh <- fmt.Sprintf("Skipping ClusterRole: %s", name)
+			continue
+		}
+
+		waitGroup.Add(1)
 		go func() {
-			err := api.Delete(ctx, name, deletePolicy)
-			if err != nil {
+			if err := api.Delete(ctx, name, deletePolicy); err != nil {
 				errorCh <- errors.Wrap(err, fmt.Sprintf("failed to delete clusterRole %s", name))
 			}
 			waitGroup.Done()
@@ -86,7 +104,18 @@ func deleteClusterRoles(clientset *kubernetes.Clientset, errorCh chan<- error) {
 	waitGroup.Wait()
 }
 
-func deleteClusterRoleBindings(clientset *kubernetes.Clientset, errorCh chan<- error) {
+var defaultClusterRoleBindings = []string{
+	"cluster-admin",
+	"docker-for-desktop-binding", // docker desktop
+	"vpnkit-controller",          // docker desktop
+}
+
+var defaultClusterRoleBindingPrefixes = []string{
+	"kubeadm:",
+	"system:",
+}
+
+func deleteClusterRoleBindings(clientset *kubernetes.Clientset, logCh chan<- string, errorCh chan<- error) {
 	ctx, cancel := createCtx()
 	waitGroup := sync.WaitGroup{}
 
@@ -97,9 +126,14 @@ func deleteClusterRoleBindings(clientset *kubernetes.Clientset, errorCh chan<- e
 		errorCh <- errors.Wrap(err, "failed to list clusterRoleBindings")
 	}
 	for _, clusterRoleBinding := range clusterRoleBindings.Items {
-		waitGroup.Add(1)
-
 		name := clusterRoleBinding.Name
+
+		if util.Contains(defaultClusterRoleBindings, name) || util.StartsWithAny(defaultClusterRoleBindingPrefixes, name) {
+			logCh <- fmt.Sprintf("Skipping ClusterRole: %s", name)
+			continue
+		}
+
+		waitGroup.Add(1)
 		go func() {
 			err := api.Delete(ctx, name, deletePolicy)
 			if err != nil {
